+++ boa3/builtin/contract/__init__.py
@@ -25,10 +25,3 @@ def abort():
     Abort the execution of a smart contract
     """
     pass
-
-
-def exit():
-    """
-    Abort the execution of a smart contract
-    """
-    pass
