+++ boa3_test/test_sc/built_in_methods_test/Exit.py
@@ -1,6 +1,4 @@
 from boa3.builtin import public
-from boa3.builtin.contract import exit
-
 
 @public
 def main(check: bool) -> int:
